+++ RELEASE_NOTES.rst
@@ -9,6 +9,7 @@ v4.1.0
 - Updated repo documents
 - Removed deprecated RouteDecorator Init
 - Updated to allow WebSocket Request to pass query arguments
+- Leverage new ApiError class for error reporting
 
 v4.0.0
 ======